# optimize: merged phase `rewrite` and `access`, make it faster.